requirements.txt
# Core Quantum Computing
qiskit==0.45.0
qiskit-aer==0.13.0
pennylane==0.33.0
cirq==1.2.0

# Machine Learning
torch==2.2.0
torchvision==0.17.0
torch-geometric==2.4.0
tensorflow==2.15.0
scikit-learn==1.3.2

# Scientific Computing
numpy==1.26.0
scipy==1.12.0
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
networkx==3.2.1

# Reinforcement Learning
gym==0.26.2
stable-baselines3==2.2.0

# Utilities
tqdm==4.66.1
joblib==1.3.2
pyyaml==6.0.1
pytest==7.4.3
black==23.11.0
jupyter==1.0.0
jupyterlab==4.0.11
ipython==8.18.1
notebook==7.0.8

# Development
pre-commit==3.5.0
flake8==6.1.0
mypy==1.7.1
